requirements.txt
tzlocal
pandas
requests
python-dotenv
google-cloud-bigquery
Flask
gunicorn
pyarrow
google-cloud-storage
google-cloud-bigquery-storage
python-dotenv
pandas_gbq